Converting for IPs
Converting this list to be used with IP address and NOT domain names. 
This will be used on pfSense since I am ditching pihole. 

Will finish this list later. I also will be adding some other lists that I use to this list with the correct IP address. 
This is 100% going to break a ton of stuff on my network! I will continue to revise and edit as I see issues with browsing.