Added my list of lists
The file that has been added is an up-to-date list of all the lists that I am using on my DNSBL inside of pfSense. 
This will not be super helpful to anyone other than me. 

This list NOT to be used directly with pfBlockerNG or any other DNS black hole,